Add files via upload
This dataset contains the results for the participants of the RunDisney Dopey Challenge in January 2022. The Dopey Challenge includes 4 back to back races, a 5K on Thursday, 10K on Friday, Half Marathon on Saturday, and Full Marathon on Sunday. In 2022, there were 5,323 participants who completed the Dopey Challenge. This dataset includes their finish times for all four races, as well as the converted miles per hour (average pace) it was completed in.